Resolve merge conflict and moved dataset package from commons to model (discussed with min yih)
