+++ src/main/java/seedu/address/logic/commands/sale/StatsCommand.java
@@ -7,11 +7,11 @@ import java.time.Month;
 import java.time.Year;
 import java.time.ZoneId;
 
-import seedu.address.commons.dataset.DataSet;
-import seedu.address.commons.dataset.date.MonthlyCountData;
 import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.model.Model;
+import seedu.address.model.dataset.DataSet;
+import seedu.address.model.dataset.date.MonthlyCountData;
 
 /**
  * Statistics on Sales.
