+++ src/main/java/seedu/address/ui/MainWindow.java
@@ -16,12 +16,12 @@ import javafx.scene.layout.StackPane;
 import javafx.stage.Stage;
 import seedu.address.commons.core.GuiSettings;
 import seedu.address.commons.core.LogsCenter;
-import seedu.address.commons.dataset.Data;
-import seedu.address.commons.dataset.DataSet;
 import seedu.address.logic.Logic;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.logic.parser.exceptions.ParseException;
+import seedu.address.model.dataset.Data;
+import seedu.address.model.dataset.DataSet;
 
 /**
  * The Main Window. Provides the basic application layout containing
